TYPE-A: i32->i8 SpaceGroupOp rotations + i32->i16 BasisSet Miller + dead index_map (#80)
* TYPE-A: narrow SpaceGroupOp rotation i32 -> i8

Rotation-matrix entries for crystallographic symmetry operations sit in
{-1, 0, 1} for cubic groups and {-2, -1, 0, 1, 2} at worst for hexagonal
settings. i8 gives ~40x headroom and shaves 27 bytes per op, packing all
48 Fd-3m ops into ~2 cache lines (from 27).

- SymmOp / SpaceGroupOp::rotation: [[i32; 3]; 3] -> [[i8; 3]; 3]
- Arithmetic widens to i32 at method entry (rotation_i32 helper) so
  det/compose/adjugate triple products never risk overflow.
- SymmOp::from_flat retains its [i32; 9] public signature; it narrows via
  i8::try_from to preserve call-site ergonomics.
- detect::narrow_rotation bridges the i32 enumeration path (unchanged) to
  the i8 storage.
- check_grid_compatibility and real-space symmetrize_density widen reads
  via i32::from / f64::from explicitly; these are cold paths.
- symmetrize_density_g's transpose_rotation folds the i8->i32 widen into
  its existing transpose pass, so the hot loop's arithmetic stays i32.

No arithmetic semantics change; all 265+ tests pass bit-identical.

* TYPE-A: narrow BasisSet Miller indices i32 -> i16

Miller index entries are bounded by g_max = sqrt(ecut / HBAR2_OVER_2M)
divided by |b_i|. i16's range [-32 768, 32 767] covers every physically
meaningful `ecut · Ω` combination with three orders of magnitude to
spare — reaching i16 overflow would require ecut > 10^7 Ry.

- BasisSet::miller: Vec<[i32; 3]> -> Vec<[i16; 3]>
- index_map keys stay (i32, i32, i32) so that `index_of` preserves its
  i32 API — out-of-range queries continue to return `None` rather than
  panic in a try_from.
- All call sites that feed Miller entries into `miller_to_idx` or
  subtraction widen explicitly via `i32::from(..)` at the read boundary.
  Storage narrows to 2 B per component; arithmetic stays i32.
- Enumeration loop in `BasisSet::new` remains i32-driven; narrowing
  happens at the push boundary via `i16::try_from(..)`.

Affected call sites (widen-on-read):
- src/scf/grid.rs: n_max reduction + basis_to_fft wrap
- src/scf/potentials.rs: build_hamiltonian_with_v_eff Miller diff
- tests/nonlocal_symmetry.rs: V_local direct-FFT assembly
- src/basis.rs: test_miller_indices_consistency roundtrip

No arithmetic semantics change; all tests pass bit-identical.

* TYPE-A: drop dead BasisSet index_map; linear scan for test-only lookup

`BasisSet::index_map: HashMap<(i32, i32, i32), usize>` was populated on
every `BasisSet::new` call (~30 kB at n_pw = 725) but `index_of` — its
only consumer — is called exclusively from test code. SCF production
paths work in the opposite direction (`miller_to_idx` / `basis_to_fft`
on an FFT grid), never through the map.

Per TYPE proposal §1.3 option (a): delete the field and reimplement
`index_of` as an O(n_pw) linear scan over the `miller` vec. The cost
is trivial next to SCF wall time (at most ~10 calls per test run,
~725 comparisons each), and the 30 kB of per-calculation dead memory
is reclaimed in full.

- BasisSet no longer stores the HashMap field (~30 kB saved per basis).
- `index_of(n1: i32, n2: i32, n3: i32) -> Option<usize>` signature
  preserved; out-of-i16-range queries fold into `None` via `try_into`
  rather than panicking — matches the prior `HashMap::get` contract.
- Added a regression probe for the i16 overflow path in the unit tests.

All 280 tests pass bit-identical.

* TYPE-A: archive TYPE proposal to completed/; update INDEX

Phase A implementation landed via the preceding three commits. Add
before/after bench numbers and per-item outcome summary to the proposal;
move to proposals/completed/ and note §2.3 (KB form-factor f32 storage)
still pending EM decision post-ITEV.

+++ src/scf/potentials.rs
@@ -154,9 +154,12 @@ pub(crate) fn build_hamiltonian_with_v_eff(
     let miller_idx = basis.miller_indices();
     for i in 0..n {
         for j in 0..n {
-            let dn1 = miller_idx[i][0] - miller_idx[j][0];
-            let dn2 = miller_idx[i][1] - miller_idx[j][1];
-            let dn3 = miller_idx[i][2] - miller_idx[j][2];
+            // Miller entries are `i16` (TYPE-A). Widen to `i32` before
+            // subtraction so differences cannot overflow even at the
+            // i16 boundary; `miller_to_idx` takes `i32` natively.
+            let dn1 = i32::from(miller_idx[i][0]) - i32::from(miller_idx[j][0]);
+            let dn2 = i32::from(miller_idx[i][1]) - i32::from(miller_idx[j][1]);
+            let dn3 = i32::from(miller_idx[i][2]) - i32::from(miller_idx[j][2]);
             let fft_idx = miller_to_idx(grid_dims, dn1, dn2, dn3);
             h[(i, j)] += v_eff_fft[fft_idx];
         }
